moved the createUser function from the registrationDAO to the userDAO and deleted the registrationDAO
